fixed help command and fixed econ command structure
