+++ commands/econ_useItem.js
@@ -1,0 +1,26 @@
+const config = require('../config.json'); // basic load of config file
+const fs = require('fs'); // File System for JS
+const pglibrary = require("../libraryfunctions.js");
+const item_handler = require('../points/item_handler');
+const points_manager = require('../points/manager');
+const {MessageEmbed, Message, MessageActionRow, MessageButton} = require('discord.js');
+module.exports = {
+    name: 'use-item',
+    description: 'Use an Item in your inventory',
+    args: 'Arguments: Item Name - For Spaces use Underscores',
+    active: true,
+    async execute(message, args, bot){
+        if(args[0]){
+            itemName = args[0].replace('_, " ');
+            item = item_handler.fetchItem(itemName, true);
+            err = points_manager.useItem(message.author.id, item);
+            if(err){
+                message.channel.send(err);
+                return;
+            } else{
+                message.channel.send(`<@${message.author.id}>, You have successfully used the Item: ${item.name}`);
+                return;
+            }
+        }
+    }
+}
